% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_anova_simulation.R
\name{anova_dat}
\alias{anova_dat}
\title{Modification of the ANOVA_power function from the Superpower package in order to obtain the simulated values
from which power calculations are performed.}
\usage{
anova_dat(
  design_result,
  alpha_level = Superpower_options("alpha_level"),
  correction = Superpower_options("correction"),
  p_adjust = "none",
  nsims = 1000,
  seed = NULL,
  verbose = Superpower_options("verbose"),
  emm = Superpower_options("emm"),
  emm_model = Superpower_options("emm_model"),
  contrast_type = Superpower_options("contrast_type"),
  emm_p_adjust = "none",
  emm_comp = NULL
)
}
\arguments{
\item{design_result}{Output from the ANOVA_design function}

\item{nsims}{Number of simulations to perform}
}
\description{
Modification of the ANOVA_power function from the Superpower package in order to obtain the simulated values
from which power calculations are performed.
}
