% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencovmat.R
\name{gencovariancemat}
\alias{gencovariancemat}
\title{Function that generates a covariance matrix taking as input a correlation matrix and a standard deviation matrix or value.}
\usage{
gencovariancemat(
  correlation_matrix,
  sd_matrix,
  withinf,
  label_list = NULL,
  nlfA,
  nlfB
)
}
\arguments{
\item{correlation_matrix}{Matrix - Expected correlation between combinations of factor levels}

\item{sd_matrix}{Numeric or matrix - Standard deviation value or matrix of standard deviation values for combinations of factor levels.}

\item{withinf}{Character- Factor for which measurements are repeated, options are NULL, "fA", "fB" and "both". If NULL (default) independent measurements will be considered.}

\item{label_list}{List length 2 - Names of factor levels}

\item{nlfA}{Integer - number of levels of factor A}

\item{nlfB}{Integer - number of levels of factor B}
}
\value{
Covariance matrix
}
\description{
May be run independently or internally as part of 'calculate_mean_matrix'.
}
