% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencovmat.R
\name{gencovmat}
\alias{gencovmat}
\title{Internal function that generates correlation and covariance matrices.
Takes input generated by the 'calculate_mean_matrix' function and runs inside of it.}
\usage{
gencovmat(mean_matrix, sd_matrix, rho = rho, label_list, withinf, nlfA, nlfB)
}
\arguments{
\item{mean_matrix}{Matrix - cell mean value matrix}

\item{sd_matrix}{Matrix - cell sd value matrix}

\item{rho}{Numeric - correlation of within factors}

\item{label_list}{List length 2 - Names of factor levels}

\item{nlfA}{Integer - number of levels of factor A}

\item{nlfB}{Integer - number of levels of factor B}
}
\value{
list length 2 with covariance and correlation matrices
}
\description{
Internal function that generates correlation and covariance matrices.
Takes input generated by the 'calculate_mean_matrix' function and runs inside of it.
}
