% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_modelled_means_sd.R
\name{graph_twoway_assumptions}
\alias{graph_twoway_assumptions}
\title{Internal function that plots modeled cell means and standard deviations and covariance matrices.
Takes input generated by the 'calculate_mean_matrix' function and runs inside of it.}
\usage{
graph_twoway_assumptions(group_size = 100, matrices_obj)
}
\arguments{
\item{group_size}{integer - number of subjects in each group}

\item{matrices_obj}{List length 2 - Cell means and standard deviation matrices}
}
\value{
Line plot with expected mean and standard deviation for each combination of factor levels
}
\description{
Internal function that plots modeled cell means and standard deviations and covariance matrices.
Takes input generated by the 'calculate_mean_matrix' function and runs inside of it.
}
