% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mean_sd_matrices.R
\name{mean_sd_matrix}
\alias{mean_sd_matrix}
\title{Create input for simulation based two-way factorial experiments}
\usage{
mean_sd_matrix(
  refmean,
  nlfA,
  nlfB,
  fAeffect,
  fBeffect,
  groupswinteraction = NULL,
  interact = 1,
  label_list = NULL,
  sdproportional = TRUE,
  sdratio = 0.2,
  endincrement = FALSE
)
}
\arguments{
\item{refmean}{Expected mean for lowest level of both factors A and B}

\item{nlfA}{Number of levels of factor A}

\item{nlfB}{Number of levels of factor B}

\item{fAeffect}{Multiple by which the refmean is modified when going from one level to the next of factor A when endincrement is FALSE (default), or multiple by which the last level of factor A is modified with respect to refmean when endincrement is TRUE}

\item{fBeffect}{Multiple by which the refmean is modified when going from one level to the next of factor B when endincrement is FALSE (default), or multiple by which the last level of factor B is modified with respect to refmean when endincrement is TRUE}

\item{groupswinteraction}{Combination of levels from factors A and B in which interaction is expected}

\item{interact}{Value by which the mean from cell or cells indicated in groupswinteraction is multiplied after it has been calculated accordingly to fAeffect and fBeffect}

\item{label_list}{List that contains vectors with the names of the factor levels. The objects in this list should be named as the factors. The use of this option is encouraged as these names are inherited to ANOVA_design.}

\item{sdproportional}{Whether the standard deviation for each combination of factor levels is a proportion of the respective factor level combination mean, defaults to TRUE}

\item{sdratio}{Value by which the expected mean value of a factor level combination is multiplied to obtain the respective standard deviation, defaults to 0.2.}

\item{endincrement}{Determines if the multiples provided in fAeffect and fBeffect refer to level to level changes (default) or change between first and last levels.}
}
\description{
The mean_sd_matrix function is intended as a companion function to the Superpower package ANOVA_design function,
followed by the Superpower ANOVA_power function.
We were motivated by sample size calculation for two-way factorial designs with a,b,...,i levels of factor A and
a,b,...,j levels of factor B in which the mean outcome value for replicates of cell A=a and B=a are known.
Furthermore, there is an estimate of the change in level mean for each of the factors. Finally, some level
combinations are expected to present interaction.
}
\details{
With user provided reference mean (mean for A=a and B=a), the expected change for each factor from one level to
the next (or from the first to last level) and the number of levels in each factor, the function creates a
matrix of expected means and a matrix of expected standard deviations with a cell for each combination of levels
i and j.

Aditionally, if the user specifies factor level combinations which are expected to present interaction and its
magnitude with respect to the reference mean, the expected change in the cell means will be incorporated to the
aforementioned matrices.
}
\examples{
refmean <- 1
treatgroups <- 4
timepoints <- 5
treateff <- 1.5
timeeff <- 0.85
effects_treat_time <- mean_sd_matrix(refmean = refmean, fAeffect = treateff, fBeffect = timeeff, nlfA = treatgroups, nlfB = timepoints, label_list = list(treatment=letters[1:treatgroups], time=1:timepoints))
muvec <- as.vector(effects_treat_time$mean.mat)
sdvec <- as.vector(effects_treat_time$sd.mat)
bwdes <- Superpower::ANOVA_design("5w*4b", n=15, mu =  muvec, sd = sdvec, label_list = rev(dimnames(effects_treat_time$mean.mat)), r=0.5)
}
