% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_powercurve.R
\name{plot_powercurve}
\alias{plot_powercurve}
\title{Plots the output of exact_twoway_anova_power}
\usage{
plot_powercurve(
  exact_anova_power,
  target_power = NULL,
  title = NULL,
  target_line = TRUE
)
}
\arguments{
\item{exact_anova_power}{List that contains a data.frame with sample sizes and corresponding powers to be plotted.}

\item{target_power}{Desired power to be attained. Accepts values between 0 and 1, defaults to 0.8.}

\item{title}{Title for the graph. Defaults to 'Power curve from exact ANOVA test'}

\item{target_line}{Set to TRUE. If FALSE no target line will be drawn. Overrides target_power.}
}
\value{
Plot with power curves.
}
\description{
Creates a power curve from the output of the exact_twoway_anova_power function.
}
\examples{

refmean <- 1
treatgroups <- 4
timepoints <- 5
treateff <- 1.5
timeeff <- 0.85
cellswithinteraction <- matrix(c(rep(2,3), 3:5), 3,2) #second level of factor A interacts with 3rd, 4th and 5th level of factor B
effects_treat_time_interact <- mean_sd_matrix(refmean = refmean, nlfA = treatgroups, nlfB = timepoints, fAeffect = treateff, fBeffect = timeeff, label_list = list(treatment=letters[1:treatgroups], time=1:timepoints),
                                              groupswinteraction = cellswithinteraction, interact=1.3)
fxs <- anoveff(effects_treat_time_interact)
powerresult <- exact_twoway_anova_power(treatgroups, timepoints, fxs, 5:20)
plot_powercurve(powerresult, target_line = 0.9)

}
