% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_powercurves.R
\name{plot_powercurves}
\alias{plot_powercurves}
\title{Plots the output of test_twoway_nrange}
\usage{
plot_powercurves(
  power_over_nrange,
  target_power = NULL,
  title = NULL,
  target_line = TRUE,
  alpha_line = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{power_over_nrange}{data.frame with sample sizes and corresponding powers to be plotted}

\item{target_power}{Desired power to be attained. Accepts values between 0 and 1, defaults to 0.8.}

\item{title}{Title for the graph. Defaults to 'Power curve from exact ANOVA test'}

\item{target_line}{Set to TRUE. If FALSE no target line will be drawn. Overrides target_power.}
}
\value{
Plot with power curves.
}
\description{
Internal function, called by test_twoway_nrange, to plot power against sample size.
}
\examples{

## 'cornorm_model' is created with the calculate_mean_matrix function

nrange <- 12:20
cornorm_sim <- simulate_twoway_nrange(nrange, cornorm_model)

power_results <- test_twoway_nrange(cornorm_sim, test="rank", plot=TRUE)

}
