% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_over_groupsize.R
\name{simulate_twoway_nrange}
\alias{simulate_twoway_nrange}
\title{Simulated independent and repeated measures two-way experiments over a set of sample sizes}
\usage{
simulate_twoway_nrange(
  matrices_obj,
  nrange,
  repeated_measurements = FALSE,
  distribution = "normal",
  skewness = 1,
  inferior_limit = -Inf,
  superior_limit = Inf,
  nsims = 500
)
}
\arguments{
\item{matrices_obj}{List - Output generated by 'calculate_mean_matrix' that include cell mean and covariance matrices}

\item{nrange}{Vector - Sample sizes to be used in simulations.}

\item{repeated_measurements}{Logical - Does the design have repeated measurements. Default is false.}

\item{distribution}{Character - - Type of distribution to simulate. Only implemented for independent designs. Possible values are 'normal', 'truncated' or 'skewed'.}

\item{skewness}{Numeric - Momentum of distribution skewness}

\item{inferior_limit}{Numeric - Value of the lower bound for the truncated distribution, defaults to '-Inf'}

\item{superior_limit}{Numeric - Value of the upper bound for the truncated distribution, defaults to 'Inf'}

\item{nsims}{Integer - Number of iterations}
}
\value{
List with of data.frames of simulated outcome values under different sample sizes. Each data.frame includes factor level labels, iteration number and sample size.
}
\description{
Wrapper for both independent and repeated measures two-way simulations. A vector of defined sample sizes is simulated under the
model provided.
}
\examples{
#' refmean <- 1
treatgroups <- 4
timepoints <- 5
treateff <- 1.5
timeeff <- 0.85
## Independent design
effects_treat_time <- calculate_mean_matrix(refmean = refmean, fAeffect = treateff, fBeffect = timeeff, nlfA = treatgroups, nlfB = timepoints, label_list = list(treatment=letters[1:treatgroups], time=1:timepoints))
## Inspect plot to check if matrices correspond to design
n <- seq(from = 16, to = 24, by = 2)

## In this case, the default 'repeated_measurements', 'distribution' and options are used.
indep_simulation <- simulate_twoway_nrange(effects_treat_time, n)

}
