% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_overseveraln.R
\name{test_power_overkn}
\alias{test_power_overkn}
\title{Test simulated two-way factorial design experiments over different sample sizes.}
\usage{
test_power_overkn(data, test = "ANOVA", plot = TRUE)
}
\arguments{
\item{data}{data.frame - data.frame with modeled outcome values, factor level labels, iteration number and sample size.}

\item{test}{character - Statistical test to be applied, possible values are 'ANOVA', 'rank' and 'permutation'.}

\item{plot}{logical - Should the power curve be plotted. Default is TRUE.}
}
\value{
Data frame with power and confidence intervals for the main effects and interaction for each of the sample sizes.
Also presented in graphical form if 'plot=TRUE'.
}
\description{
Wrapper to test data simulated under independent or repeated measurements and under different outcome distributions
with different sample sizes.
}
\examples{


## In this example we simulate an independent sample design with skewed outcome
## Model was specified with the 'calculate_mean_matrix function' (see ?calculate_mean_matrix)
refmean <- 1
treatgroups <- 4
timepoints <- 5
treateff <- 1.5
timeeff <- 0.85
label_list <- list(treatment=letters[1:treatgroups], time=1:timepoints)

indep_matrix <- calculate_mean_matrix(refmean = refmean, fAeffect = treateff, fBeffect = timeeff, nlfA = treatgroups, nlfB = timepoints, label_list = label_list)

indep_skewsim <- simulate_twoway_nrange(indep_matrix, seq(3, 12, 3), distribution = "skewed", skewness = 1.8)
test_power_overkn(indep_skewsim)

}
