% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_twoway_generation.R
\name{twoway_simulation_correlated}
\alias{twoway_simulation_correlated}
\title{Simulate measurements repeated over either or both factors of a two-way design}
\usage{
twoway_simulation_correlated(group_size, matrices_obj, nsims = 500)
}
\arguments{
\item{group_size}{Integer - sample size for each cell}

\item{matrices_obj}{List - Output generated by 'calculate_mean_matrix' that include cell mean and covariance matrices}

\item{nsims}{Integer - Number of iterations}
}
\value{
data.frame with modeled outcome values, factor level labels, iteration number and sample size.
}
\description{
Both regular and internal function.
As regular function takes input generated by the 'calculate_mean_matrix' function and iteratively simulates repeated measures two-way factorial experiments.
Data are sampled from a normal distribution.
}
\details{
As internal function runs with a single iteration inside 'graph_twoway_assumptions', which in itself is inside  'calculate_mean_matrix' to generate data for the cell mean and standard deviation plot.
}
\examples{
## Repeated measures design, suppose subjects from 4 independent treatment groups measured at 5 different timepoints.

refmean <- 1
treatgroups <- 4
timepoints <- 5
treateff <- 1.5
timeeff <- 0.85
rho <- 0.8
withinf <- "fB"

effects_treat_time <- calculate_mean_matrix(refmean = refmean, fAeffect = treateff, fBeffect = timeeff, nlfA = treatgroups,  nlfB = timepoints,
rho = rho, withinf = withinf, label_list = list(treatment=letters[1:treatgroups], time=1:timepoints))

## Inspect plot to check if matrices correspond to design
n <- 20
repeatedmeasures_experiment <- twoway_simulation_correlated(group_size = n, matrices_obj = effects_treat_time)
head(repeatedmeasures_experiment, 10)

}
