% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_fromeffectsize.R
\name{exact_twoway_anova_power}
\alias{exact_twoway_anova_power}
\title{Two-way factorial ANOVA exact sample size calculation for independent samples}
\usage{
exact_twoway_anova_power(
  a,
  b,
  effect_sizes,
  n,
  alpha = 0.05,
  factor_names = NULL
)
}
\arguments{
\item{a}{Number of levels of the first factor}

\item{b}{Number of levels of the second factor}

\item{effect_sizes}{Vector of length 3. The first two elements are the effect sizes for the main effects of the first
and second factors, respectively. The third element is the interaction effect size.}

\item{n}{Number of experimental units in each group for which power (1-beta) will be calculated.}

\item{alpha}{Accepted type 1 error. Defaults to 0.05}

\item{factor_names}{Character vector of length 2. Names of the 2 factors to be evaluated. Default is to inherit names
from effect_sizes. If effect_sizes has no names and no factor_names are provided, factors will be named 'FactorA' and
'FactorB'.}
}
\value{
A list that contains the number of levels for each factor, the chosen significance level and a data.frame in which the
first column is the group sample size and the remaining three columns are the power for the main effect of the first
factor, the main effect of the second factor and their interaction, respectively.
}
\description{
This functions takes the effect sizes (Cohen's f) for two main effects and their interaction and estimates power
a range of sample sizes. The input for this function can be generated by anoveff.
}
\examples{
refmean <- 1
treatgroups <- 4
timepoints <- 5
treateff <- 1.25
timeeff <- 0.85
cellswithinteraction <- matrix(c(rep(2,3), 3:5), 3,2) #second level of factor A interacts with 3rd, 4th and 5th level of factor B
effects_treat_time_interact <- mean_sd_matrix(refmean = refmean, nlfA = treatgroups, nlfB = timepoints, fAeffect = treateff, fBeffect = timeeff, label_list = list(treatment=letters[1:treatgroups], time=1:timepoints),
                                             groupswinteraction = cellswithinteraction, interact=1.3)
fxs <- anoveff(effects_treat_time_interact)

exact_twoway_anova_power(a= treatgroups, b=timepoints, effect_sizes=fxs, n=5:20)

}
