% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencovmat.R
\name{gencovmat}
\alias{gencovmat}
\title{Internal function that generates correlation and covariance matrices.}
\usage{
gencovmat(mean_matrix, sd_matrix, rho, label_list = NULL, withinf, nlfA, nlfB)
}
\arguments{
\item{mean_matrix}{Matrix - cell mean value matrix}

\item{sd_matrix}{Matrix - cell sd value matrix}

\item{rho}{Vector length 1 or 2, or 2 by 2 matrix - Controls how the correlation and hence de covariance matrix is built.}

\item{label_list}{List length 2 - Names of factor levels}

\item{withinf}{Character- Factor for which measurements are repeated, options are NULL, "fA", "fB" and "both". If NULL (default) independent measurements will be considered.}

\item{nlfA}{Integer - number of levels of factor A}

\item{nlfB}{Integer - number of levels of factor B}
}
\value{
list length 2 with covariance and correlation matrices
}
\description{
Takes input generated by the 'calculate_mean_matrix' function and runs inside of it.
}
