% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_twoway_generation.R
\name{twoway_simulation_correlated}
\alias{twoway_simulation_correlated}
\title{Simulate measurements repeated over either or both factors of a two-way design}
\usage{
twoway_simulation_correlated(
  group_size,
  matrices_obj,
  distribution = "normal",
  shape.parameter = 0,
  inferior_limit = -Inf,
  superior_limit = Inf,
  nsims = 500
)
}
\arguments{
\item{group_size}{Integer - sample size for each cell}

\item{matrices_obj}{List - Output generated by 'calculate_mean_matrix' that include cell mean and covariance matrices}

\item{distribution}{Character - Type of distribution from which to sample, possible values are "normal", "skewed" and "truncated"}

\item{shape.parameter}{Vector - Degree of skewness in the distribution. May be a single value, hava a length equal to the number of levels of any one of the factors or a length equal to the product of the length of each factor.}

\item{inferior_limit}{Numeric - Value for which the distribution is truncated on the left. Only valid if 'distribution="truncated"'}

\item{superior_limit}{Numeric - Value for which the distribution is truncated on the right. Only valid if 'distribution="truncated"'}

\item{nsims}{Integer - Number of iterations}
}
\value{
Dataframe with simulated outcome values, factor level labels and iteration number.
}
\description{
Both regular and internal function.
As regular function takes input generated by the 'calculate_mean_matrix' function and iteratively simulates repeated measures two-way factorial experiments.
Data are sampled from a normal, skewed normal or truncated normal distribution.
}
\details{
As internal function runs with a single iteration inside 'graph_twoway_assumptions', which in itself is inside  'calculate_mean_matrix' to generate data for the cell mean and standard deviation plot.
}
\examples{
## Repeated measures design, suppose subjects from 4 independent treatment groups measured at 5 different timepoints.

refmean <- 1
treatgroups <- 4
timepoints <- 5
treateff <- 1.5
timeeff <- 0.85
rho <- 0.8
withinf <- "fB"

effects_treat_time <- calculate_mean_matrix(refmean = refmean, fAeffect = treateff, fBeffect = timeeff, nlfA = treatgroups,  nlfB = timepoints,
rho = rho, withinf = withinf, label_list = list(treatment=letters[1:treatgroups], time=1:timepoints))

## Inspect plot to check if matrices correspond to design
n <- 20
repeatedmeasures_experiment <- twoway_simulation_correlated(group_size = n, matrices_obj = effects_treat_time)
head(repeatedmeasures_experiment, 10)

}
