% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_twoway_generation.R
\name{twoway_simulation_independent}
\alias{twoway_simulation_independent}
\title{Simulate independent measurements in a two-way factorial design}
\usage{
twoway_simulation_independent(
  group_size,
  matrices_obj,
  distribution = "normal",
  skewness = 1,
  inferior_limit = -Inf,
  superior_limit = Inf,
  balanced = TRUE,
  nsims = 500
)
}
\arguments{
\item{group_size}{Integer - sample size for each cell}

\item{matrices_obj}{List - Output generated by 'calculate_mean_matrix' that include cell mean and standard deviation matrices}

\item{distribution}{Character - Type of distribution to simulate. Possible values are 'normal', 'truncated' or 'skewed'.}

\item{skewness}{Numeric - Momentum of distribution skewness}

\item{inferior_limit}{Numeric - Value of the lower bound for the truncated distribution, defaults to '-Inf'}

\item{superior_limit}{Numeric - Value of the upper bound for the truncated distribution, defaults to 'Inf'}

\item{nsims}{Integer - Number of iterations}
}
\value{
data.frame with modeled outcome values, factor level labels, iteration number and sample size.
}
\description{
Both regular and internal function.
As regular function takes input generated by the 'calculate_mean_matrix' function and iteratively simulates independent measures two-way factorial experiments.
Outcome may be normally distributed, truncated normally distributed or skewed.
}
\details{
As internal function runs with a single iteration inside 'graph_twoway_assumptions', which in itself is inside  'calculate_mean_matrix' to generate data for the cell mean and standard deviation plot.
}
\examples{
refmean <- 1
treatgroups <- 4
timepoints <- 5
treateff <- 1.5
timeeff <- 0.85
## Independent design
effects_treat_time <- calculate_mean_matrix(refmean = refmean, fAeffect = treateff, fBeffect = timeeff, nlfA = treatgroups, nlfB = timepoints, label_list = list(treatment=letters[1:treatgroups], time=1:timepoints))
## Inspect plot to check if matrices correspond to design
n <- 20
independent_experiment <- twoway_simulation_independent(group_size = n, matrices_obj = effects_treat_time)
head(independent_experiment, 10)

}
